Rename ScaleRenderer to DefaultScaleRenderer
